refactor(economy): split cog into _economy helper modules
